backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
openai
